unbuging

+++ database/migrations/2023_03_01_083906_create_post_user_likes_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('post_user_likes', function (Blueprint $table) {
             $table->id();
-            $table->unsignedSmallInteger('post_id');
-            $table->unsignedSmallInteger('user_id');
+            $table->unsignedBigInteger('post_id');
+            $table->unsignedBigInteger('user_id');
 
             $table->index('post_id' , 'pul_post_idx');
             $table->index('user_id' , 'pul_user_idx');
